web_app/requirements.txt
Flask
psycopg2-binary
gunicorn
requests
Flask-Mail
passlib[bcrypt]
PyJWT
flasgger==0.9.7.1
